Fixing conditional statement checking for zeros.
line 247: if lastFloorHeight == 0.0 will cause errors due to lack of
floating point precision. Replaced with dedicated is_near_zero function.
